First commit of new angular correlation option
This is intended to allow users to specify which angular distribution
(Z_0, Z_2, or Z_4, see the paper from Smith et al.) to create and for
which two gamma energies. The command for this is
"/DetSys/gun/AngularCorrelation" and takes three arguments, first one is
the type (0, 2, or 4), second is the lower/draining energy, third is the
higher/feeding energy.

Doesn't work yet!